+++ dash-renderer/src/APIController.react.js
@@ -1,5 +1,5 @@
 import {connect} from 'react-redux';
-import {contains, isEmpty, isNil} from 'ramda';
+import {includes, isEmpty, isNil} from 'ramda';
 import React, {Component} from 'react';
 import PropTypes from 'prop-types';
 import TreeContainer from './TreeContainer';
@@ -10,6 +10,7 @@ import {
     hydrateInitialOutputs,
     setLayout,
 } from './actions/index';
+import {applyPersistence} from './persistence';
 import apiThunk from './actions/api';
 import {getAppState} from './reducers/constants';
 import {STATUS} from './constants/constants';
@@ -48,7 +49,11 @@ class UnconnectedContainer extends Component {
             dispatch(apiThunk('_dash-layout', 'GET', 'layoutRequest'));
         } else if (layoutRequest.status === STATUS.OK) {
             if (isEmpty(layout)) {
-                dispatch(setLayout(layoutRequest.content));
+                const finalLayout = applyPersistence(
+                    layoutRequest.content,
+                    dispatch
+                );
+                dispatch(setLayout(finalLayout));
             } else if (isNil(paths)) {
                 dispatch(computePaths({subTree: layout, startingPath: []}));
             }
@@ -102,13 +107,13 @@ class UnconnectedContainer extends Component {
 
         if (
             layoutRequest.status &&
-            !contains(layoutRequest.status, [STATUS.OK, 'loading'])
+            !includes(layoutRequest.status, [STATUS.OK, 'loading'])
         ) {
             return <div className="_dash-error">Error loading layout</div>;
         } else if (
             errorLoading ||
             (dependenciesRequest.status &&
-                !contains(dependenciesRequest.status, [STATUS.OK, 'loading']))
+                !includes(dependenciesRequest.status, [STATUS.OK, 'loading']))
         ) {
             return (
                 <div className="_dash-error">Error loading dependencies</div>